requirements.txt
neo4j==5.14.1
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
networkx==3.2.1 